Fix "the dialect is not supported" error
This makes it so that, if the provided dialect is a supported dialect (eg. postgres) and requiring the dialect throws an error (eg. the 'pg' package not being installed), the actual underlying error and its accompanying stack trace are thrown to the caller, rather than stringifying the error (losing the stack) and reporting the wrong error.