Docs: Updating getting started with Your first query (#6315)
Also correct the section about promises to sound more nice.

Also `fieldName` and `freezeTableName` is nice features, but not something you need in a getting started guide. It just makes it look more complicated.

I even think that the `pool:` option in "Setting up a connection" should be removed, because who needs that anyways when they are first getting started? I could make new people think that they need it.